Commit Corrections MVC Hache

+++ src/main/java/com/example/zeldasae/modele/Boss.java
@@ -8,7 +8,7 @@ import javafx.util.Duration;
 
 public class Boss extends Ennemi {
 
-    private final int cooldownAttaqueDistance = 3;
+    private final double cooldownAttaqueDistance = 3;
     private boolean peutAttaquerDistance;
 
     public Boss(int x, int y, int width, int height, int column, int rows, BFS bfs) {
